Improve performance by removing code
